Add new command - remark
This command allows users of the AddressBook application to add optional
remarks to people in their address book and edit it if required. The
command should have the following format:
remark INDEX r/REMARK
INDEX: Refers to the position of the person in the currently displayed
list on the GUI.
REMARK: Is the note you wish to attach to their entry.